Repace Luxon with date-fns
While better than Moment, Luxon's API is still verbose and sometimes
inconvenient. For example, `Interval.fromDateTimes` fails to calculate
the diff if of two dates if the second date is before the first.

I've become a fan of date-fns. It's not object-oriented, just a
collection of simple functions that work on Date, date strings, and
even numeric timestamps.

Unfortunately, the npm package does not contain dist files, so
download the min.js from a CDN in a build script for now.

[1] https://moment.github.io/luxon/
[2] https://date-fns.org/