Load records from Strava steams
Store records in a separate directory in the db. To keep files small
and quick to parse, store records as an object of arrays. This avoids
replicating the keys for each element. Restore records objects after
reading these files.

Remove the `records` field from activities to avoid keeping the records
for all activities in memory. Load only the records for the selected
activity.

Provide a model function `getActivityRecords` in the to do so.
Add an event `get-records` to request records from the main thread,
and an event `records` to provide records to the renderer.